Rename Wigwam assets to OG
