+++ src/app/components/blocks/AddAccountModal.tsx
@@ -26,7 +26,7 @@ import Button from "app/components/elements/Button";
 import BackButton from "app/components/elements/BackButton";
 import ScrollAreaContainer from "app/components/elements/ScrollAreaContainer";
 import AddAccountSteps from "app/components/blocks/AddAccountSteps";
-import { ReactComponent as WigwamIcon } from "app/icons/Wigwam.svg";
+import { ReactComponent as OGIcon } from "app/icons/OG.svg";
 
 const AddAccountModal = memo(() => {
   const [accModalOpened, setAccModalOpened] = useAtom(addAccountModalAtom);
@@ -140,7 +140,8 @@ const AddAccountModal = memo(() => {
         >
           <OnMount handle={handleContentMount} />
 
-          <WigwamIcon
+          {/* TODO REBRANDING: replace logo artwork */}
+          <OGIcon
             className={classNames(
               "w-16 h-auto",
               "absolute",
