+++ src/app/components/screens/Unlock.tsx
@@ -7,7 +7,7 @@ import { Page } from "app/nav";
 import { currentProfileAtom } from "app/atoms";
 import { openInTab } from "app/helpers";
 import { ReactComponent as ChangeProfileIcon } from "app/icons/change-profile.svg";
-import { ReactComponent as WigwamIcon } from "app/icons/WigwamTitle.svg";
+import { ReactComponent as OGIcon } from "app/icons/OGTitle.svg";
 
 import BoardingPageLayout from "../layouts/BoardingPageLayout";
 import PopupLayout from "../layouts/PopupLayout";
@@ -62,7 +62,8 @@ const Unlock: FC<UnlockProps> = ({ isApproval }) => {
             "flex items-center",
           )}
         >
-          <WigwamIcon className={classNames("h-[1.375rem]", "w-auto mr-3")} />
+          {/* TODO REBRANDING: replace logo artwork */}
+          <OGIcon className={classNames("h-[1.375rem]", "w-auto mr-3")} />
         </div>
       )}
     </>
